@@ -1,13 +1,24 @@
-const express = require("express");
+import express from "express";
+import path from "path";
+import mealsRouter from "./api/meals.js";
+import reservationsRouter from "./api/reservations.js";
+import reviewsRouter from "./api/reviews.js";
+import cors from "cors";
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+import knex from "knex";
+import { error } from "console";
+
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
 const app = express();
 const router = express.Router();
-const path = require("path");
 
-const mealsRouter = require("./api/meals");
 const buildPath = path.join(__dirname, "../../dist");
 const port = process.env.PORT || 3000;
-const cors = require("cors");
-const knex = require("./database");
+
 
 // For week4 no need to look into this!
 // Serve the built client html
@@ -20,78 +31,9 @@ app.use(express.json());
 
 app.use(cors());
 
-router.use("/meals", mealsRouter);
-
-// Respond with all meals in the future (relative to the when datetime)
-app.get("/future-meals", async(req, res) => {
-  try {
-    const futureMeals = await knex("Meal")
-      .select()
-      .where('when', '>', new Date());
-    if (futureMeals.length === 0) {
-      res.status(404).send("The future meals is empty");
-      return;
-    }
-    res.status(200).json(futureMeals);
-  } catch(error) {
-      console.error(error);
-      res.status(500).send("Something went wrong");
-  }
-});
-
-// Respond with all meals in the past (relative to the when datetime)
-app.get("/past-meals", async (req, res) => {
-  try {
-    const date = new Date();
-    const pastMeals = await knex("Meal")
-      .select()
-      .where("when", "<", date);
-    res.status(200).json(pastMeals);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send("Something went wrong");
-  }
-});
-
-//Respond with all meals sorted by ID
-app.get("/all-meals", async (req, res) => {
-  try {
-    const allMeals = await knex("Meal")
-      .select()
-      .orderBy('id');
-    res.status(200).json(allMeals);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send("Something went wrong");
-  }
-});
-
-//	Respond with the first meal (meaning with the minimum id)
-app.get("/first-meal", async (req, res) => {
-  try {
-    const firstMeal = await knex("Meal").select().orderBy("id").first();
-
-    res.status(200).json(firstMeal);
-
-  } catch (error) {
-    console.error(error);
-    res.status(500).send("Something went wrong");
-  }
-});
-
-//	Respond with the last meal (meaning with the maximum id)
-app.get("/last-meal", async (req, res) => {
-  try {
-
-    const firstMeal = await knex("Meal").select().orderBy("id", 'desc').first();
-  res.status(200).json(lastMeal);
-
-  } catch (error) {
-    console.error(error);
-    res.status(500).send("Something went wrong");
-  }
-});
-
+app.use("/api/meals", mealsRouter);
+app.use("/api/reservations", reservationsRouter);
+app.use("/api/reviews", reviewsRouter);
 
 if (process.env.API_PATH) {
   app.use(process.env.API_PATH, router);
@@ -104,13 +46,10 @@ app.use("*", (req, res) => {
   res.sendFile(path.join(`${buildPath}/index.html`));
 });
 
-<<<<<<< Updated upstream
-module.exports = app;
-=======
 // Respond with all meals in the future (relative to the when datetime)
 app.get("/future-meals", async (req, res) => {
   try {
-    const futureMeals = await knex("Meal")
+    const futureMeals = await knex("meal")
       .select()
       .where("when", ">", new Date());
     if (futureMeals.length === 0) {
@@ -128,7 +67,7 @@ app.get("/future-meals", async (req, res) => {
 app.get("/past-meals", async (req, res) => {
   try {
     const date = new Date();
-    const pastMeals = await knex("Meal").select().where("when", "<", date);
+    const pastMeals = await knex("meal").select().where("when", "<", date);
     res.status(200).json(pastMeals);
   } catch (error) {
     console.error(error);
@@ -139,9 +78,7 @@ app.get("/past-meals", async (req, res) => {
 //Respond with all meals sorted by ID
 app.get("/all-meals", async (req, res) => {
   try {
-    const allMeals = await knex("Meal")
-      .select()
-      .orderBy('id');
+    const allMeals = await knex("meal").select().orderBy("id");
     res.status(200).json(allMeals);
   } catch (error) {
     console.error(error);
@@ -152,8 +89,10 @@ app.get("/all-meals", async (req, res) => {
 //	Respond with the first meal (meaning with the minimum id)
 app.get("/first-meal", async (req, res) => {
   try {
-    const firstMeal = await knex("Meal").select().orderBy("id").first();
+    const firstMeal = await knex("meal").select().orderBy("id").first();
+
     res.status(200).json(firstMeal);
+
   } catch (error) {
     console.error(error);
     res.status(500).send("Something went wrong");
@@ -163,13 +102,18 @@ app.get("/first-meal", async (req, res) => {
 //	Respond with the last meal (meaning with the maximum id)
 app.get("/last-meal", async (req, res) => {
   try {
-    const lastMeal = await knex("Meal").select().orderBy("id", "desc").first();
+
+    const lastMeal = await knex("meal").select().orderBy("id", "desc").first();
     res.status(200).json(lastMeal);
+
   } catch (error) {
     console.error(error);
     res.status(500).send("Something went wrong");
   }
 });
 
+
+module.exports = app;
+
 export default app;
->>>>>>> Stashed changes
+
